+++ src/components/map/Sidebar/SearchList.tsx
@@ -58,7 +58,8 @@ import LoadingImg from "../../../../public/1Cademy_Loading_Dots.gif";
 // import TagSearch from "../../../../PublicComps/TagSearch/TagSearch";
 import { useTagsTreeView } from "../../../hooks/useTagsTreeView";
 import shortenNumber from "../../../lib/utils/shortenNumber";
-import { SortDirection, SortValues } from "../../../noteBookTypes";
+import { SortDirection, SortValues } from "../../../nodeBookTypes";
+// import { SortDirection, SortValues } from "../../../noteBookTypes";
 import { NodeType } from "../../../types";
 import { Editor } from "../../Editor";
 import NodeTypeIcon from "../../NodeTypeIcon2";
