+++ src/pages/dashboard.tsx
@@ -69,8 +69,8 @@ import {
 } from "../lib/utils/Map.utils";
 import { newId } from "../lib/utils/newid";
 import { buildFullNodes, getNodes, getUserNodeChanges } from "../lib/utils/nodesSyncronization.utils";
-import { ChoosingType, UserNodes, UserNodesData } from "../nodeBookTypes";
-import { ClusterNodes, FullNodeData } from "../noteBookTypes";
+import { ChoosingType, ClusterNodes, FullNodeData, UserNodes, UserNodesData } from "../nodeBookTypes";
+// import { ClusterNodes, FullNodeData } from "../noteBookTypes";
 import { NodeType } from "../types";
 
 type DashboardProps = {};
